シャンデリアの近くでは、懐中電灯をつけていなくてもSAN値が減らないようにした。 エネミーがプレイヤーを見失った後、経路探査したあとに壁を貫通しないようにした。(できてるはず)

+++ GameTemplate/Game/FlashLightConstant.h
@@ -33,7 +33,7 @@ namespace nsHikageri
 
 		namespace nsAbilityClairvoyanceConstant
 		{
-			static const int INI_CRAIRVOYANCE_BATTERY_COST = 0.05f;	//クレアボヤンスで消費するバッテリー(発動中毎フレーム)
+			static const float INI_CRAIRVOYANCE_BATTERY_COST = 0.05f;	//クレアボヤンスで消費するバッテリー(発動中毎フレーム)
 		}
 
 		namespace nsAbilityMedousaEyeConstant
